Move the payment provider settings into their own file
src/shared/db/settings.ts assembles the whole settings namespace, and it
also held every payment provider decision: which provider takes new sales,
which provider owns the payments that already exist, and the stored Stripe,
Square, and SumUp credentials. That put one large statement and three
credential namespaces in the file whose job is assembly, and it kept the
file at 593 lines, well above the 400-line target.

The provider code now lives in src/shared/db/settings/payment-provider.ts.
It exports the getters to merge onto `settings` and the writers to spread
onto `settings.update`, the same shape `stringAccessors` already uses. The
behaviour is unchanged: the same statement, the same snapshot writes, and
the same names on the namespace.

withProperties moves to src/shared/db/settings/namespace.ts, because both
files now merge parts that build their own getters. A spread cannot do this
work. A spread calls every getter once, at assembly time.

The tests move to the mirror path the mutation runner reads:
test/shared/db/settings/payment-provider/activation.test.ts holds the
provider choice, and credentials.test.ts holds the stored credentials. Ten
test names in the moved credentials suite named helpers that the settings
namespace replaced, so they now say what the behaviour is.

settings.ts is 373 lines.
