+++ src/shared/db/settings.ts
@@ -12,16 +12,12 @@
  * `./settings/`, each file named for the job it does.
  */
 
-import * as v from "valibot";
-
 import type { AddressLookupSetting } from "#shared/address-lookup/types.ts";
 import { isAddressLookupSetting } from "#shared/address-lookup/types.ts";
 import {
   type EnabledFeatures,
   parseEnabledFeatures,
 } from "#shared/admin-features.ts";
-import { encrypt } from "#shared/crypto/encryption.ts";
-import { executeWithoutCacheInvalidation } from "#shared/db/client.ts";
 import {
   boolUpdate,
   rawUpdate,
@@ -39,20 +35,18 @@ import {
   getCurrentSettingsVersion,
   prefetchVersion,
 } from "#shared/db/settings/cache.ts";
-import { keyModeOf } from "#shared/db/settings/constants.ts";
 import { withCurrentTask } from "#shared/db/settings/current-task.ts";
 import { invalidateCache, loadKeys } from "#shared/db/settings/load.ts";
+import { withProperties } from "#shared/db/settings/namespace.ts";
 import { updateUserPassword } from "#shared/db/settings/password.ts";
+import { paymentProviderAccessors } from "#shared/db/settings/payment-provider.ts";
 import {
-  deleteRaw,
   encryptedUpdate,
   getRawCached,
   plaintextUpdate,
-  syncStoredSetting,
   writeEncrypted,
   writeOrDelete,
   writeRaw,
-  writeRawBatch,
 } from "#shared/db/settings/raw-writes.ts";
 import {
   clearSetupCompleteCache,
@@ -73,7 +67,6 @@ import {
   parseListingDefaults,
   serializeListingDefaults,
 } from "#shared/listing-defaults.ts";
-import { PAYMENT_PROVIDER_IDS } from "#shared/payment-providers.ts";
 import { CONFIG_KEYS } from "#shared/settings/keys.ts";
 import { EMAIL_BODY_KEYS } from "#shared/settings/registry.ts";
 import {
@@ -85,16 +78,10 @@ import type { TableLayout } from "#shared/tables/layout.ts";
 import type {
   EmailTemplateFormat,
   EmailTemplateType,
-  PaymentProviderSetting,
-  PaymentProviderType,
   SuperuserChoice,
   Theme,
 } from "#shared/types.ts";
-import {
-  isPaymentProvider,
-  isSuperuserChoice,
-  PaymentProviderSettingSchema,
-} from "#shared/types.ts";
+import { isSuperuserChoice } from "#shared/types.ts";
 import { appleWallet } from "#shared/wallets/apple-wallet-settings.ts";
 import { googleWallet } from "#shared/wallets/google-wallet-settings.ts";
 import type { EmailContent } from "#templates/email/shared.ts";
@@ -108,98 +95,6 @@ export {
   getCurrentSettingsVersion,
   SetupAlreadyCompleteError,
   SNAPSHOT_KEYS,
-};
-
-/**
- * Copy property descriptors (preserving getters) from `props` onto `target`.
- * A spread would eagerly evaluate the getters instead.
- */
-const withProperties = <T extends object, P extends object>(
-  target: T,
-  props: P,
-): T & P => {
-  Object.defineProperties(target, Object.getOwnPropertyDescriptors(props));
-  return target as T & P;
-};
-
-/** The card-provider getters shared by Stripe and SumUp: whether a secret key
- * is set, and whether that key is a test or live key. `keyName` is the setting
- * the provider's secret is stored under. */
-const providerKeyStatus = (keyName: "stripe_secret_key" | "sumup_api_key") => ({
-  get hasKey(): boolean {
-    return snap(keyName) !== "";
-  },
-  get keyMode(): "test" | "live" | null {
-    return keyModeOf(snap(keyName));
-  },
-});
-
-/** Store the new-sales choice and the provider for existing payments together. */
-const setPaymentProviderSnapshot = (active: PaymentProviderSetting): void => {
-  data.payment_provider = active === "none" ? null : active;
-  data.payment_provider_setting = active;
-};
-
-const syncReturnedPaymentProvider = (active: PaymentProviderSetting): void => {
-  syncStoredSetting(CONFIG_KEYS.PAYMENT_PROVIDER, (values) =>
-    values.set(CONFIG_KEYS.PAYMENT_PROVIDER, active),
-  );
-  setPaymentProviderSnapshot(active);
-};
-
-const syncLastActivePaymentProvider = (provider: string): void => {
-  syncStoredSetting(CONFIG_KEYS.LAST_ACTIVE_PAYMENT_PROVIDER, (values) =>
-    values.set(CONFIG_KEYS.LAST_ACTIVE_PAYMENT_PROVIDER, provider),
-  );
-  data.last_active_payment_provider = provider;
-};
-
-const changePaymentProvider = async (
-  kind: "activate" | "credentials" | "disable" | "recover",
-  provider?: PaymentProviderType,
-  first = false,
-): Promise<void> => {
-  if (kind !== "disable" && (!provider || !isPaymentProvider(provider))) {
-    throw new Error(`Invalid payment provider: ${provider}`);
-  }
-  const chosen = provider ?? "";
-  const firstCredential = kind === "credentials" && first ? 1 : 0;
-  const sql = `INSERT INTO settings (key, value) SELECT key, value FROM (
-SELECT 'payment_provider' AS key, CASE WHEN ? IN ('disable', 'recover') THEN 'none' WHEN ? = 'credentials' THEN CASE WHEN (SELECT value FROM settings WHERE key = 'payment_provider') = ? OR (? = 1 AND NOT EXISTS (SELECT 1 FROM settings WHERE key = 'payment_provider')) THEN ? ELSE COALESCE((SELECT value FROM settings WHERE key = 'payment_provider'), 'none') END ELSE ? END AS value
-UNION ALL SELECT 'last_active_payment_provider', CASE WHEN ? = 'disable' THEN CASE WHEN (SELECT value FROM settings WHERE key = 'payment_provider') IN (${PAYMENT_PROVIDER_IDS.map(() => "?").join(", ")}) THEN (SELECT value FROM settings WHERE key = 'payment_provider') ELSE COALESCE((SELECT value FROM settings WHERE key = 'last_active_payment_provider'), '') END WHEN ? = 'credentials' THEN CASE WHEN (SELECT value FROM settings WHERE key = 'payment_provider') = ? OR (? = 1 AND NOT EXISTS (SELECT 1 FROM settings WHERE key = 'payment_provider')) THEN ? ELSE COALESCE((SELECT value FROM settings WHERE key = 'last_active_payment_provider'), '') END ELSE ? END
-UNION ALL SELECT 'settings_version', CAST(COALESCE((SELECT value FROM settings WHERE key = 'settings_version'), '0') AS INTEGER) + 1) WHERE (? <> 'recover' OR (COALESCE((SELECT value FROM settings WHERE key = 'payment_provider'), 'none') = 'none' AND COALESCE((SELECT value FROM settings WHERE key = 'last_active_payment_provider'), '') = '' AND EXISTS (SELECT 1 FROM settings WHERE key = CASE ? WHEN 'stripe' THEN ? WHEN 'square' THEN ? WHEN 'sumup' THEN ? END AND value <> '')))
-AND (? <> 'activate' OR NOT (COALESCE((SELECT value FROM settings WHERE key = 'payment_provider'), 'none') = 'none' AND COALESCE((SELECT value FROM settings WHERE key = 'last_active_payment_provider'), '') = '' AND (SELECT COUNT(*) FROM settings WHERE key IN (?, ?, ?) AND value <> '') > 1))
-ON CONFLICT(key) DO UPDATE SET value = excluded.value RETURNING key, value`;
-  const credentialKeys = [
-    CONFIG_KEYS.STRIPE_SECRET_KEY,
-    CONFIG_KEYS.SQUARE_ACCESS_TOKEN,
-    CONFIG_KEYS.SUMUP_API_KEY,
-  ];
-  const args = [
-    [kind, kind, chosen, firstCredential, chosen, chosen, kind],
-    PAYMENT_PROVIDER_IDS,
-    [kind, chosen, firstCredential, chosen, chosen, kind, chosen],
-    credentialKeys,
-    [kind],
-    credentialKeys,
-  ].flat();
-  const result = await executeWithoutCacheInvalidation(sql, args);
-  const rejection =
-    kind === "recover"
-      ? "Payment provider recovery is no longer available"
-      : "Choose the provider for existing payments before enabling new sales";
-  if (result.rows.length === 0) {
-    throw new Error(rejection);
-  }
-  const values = Object.fromEntries(
-    result.rows.map((row) => [String(row.key), String(row.value)]),
-  );
-  syncReturnedPaymentProvider(
-    v.parse(PaymentProviderSettingSchema, values[CONFIG_KEYS.PAYMENT_PROVIDER]),
-  );
-  syncLastActivePaymentProvider(
-    v.parse(v.string(), values[CONFIG_KEYS.LAST_ACTIVE_PAYMENT_PROVIDER]),
-  );
 };
 
 const settingsBase = {
@@ -304,13 +199,6 @@ const settingsBase = {
   // --- Google Wallet ---
   googleWallet: googleWallet.createReadSettings(snap as (k: string) => string),
   invalidateCache,
-  /** Keeps existing-payment work available while new sales are off. */
-  get lastActivePaymentProvider(): PaymentProviderType | null {
-    const value = snap("last_active_payment_provider");
-    if (value === "") return null;
-    if (isPaymentProvider(value)) return value;
-    throw new Error(`Invalid last_active_payment_provider setting: ${value}`);
-  },
   get listingColumnLayout(): TableLayout<ListingColumnKey> {
     return configurableTableLayouts.listing.parse(snap("listing_column_order"));
   },
@@ -324,12 +212,6 @@ const settingsBase = {
   },
   get orphanPurgeRetention(): string {
     return snap("orphan_purge_retention");
-  },
-  get paymentProvider(): PaymentProviderType | null {
-    return snap("payment_provider");
-  },
-  get paymentProviderSetting(): PaymentProviderSetting | null {
-    return snap("payment_provider_setting");
   },
   get phonePrefix(): string {
     return snap("phone_prefix");
@@ -370,54 +252,6 @@ const settingsBase = {
     },
   },
 
-  // --- Square ---
-  square: {
-    get accessToken(): string {
-      return snap("square_access_token");
-    },
-    get hasToken(): boolean {
-      return snap("square_access_token") !== "";
-    },
-    get locationId(): string {
-      return snap("square_location_id");
-    },
-    get sandbox(): boolean {
-      return snap("square_sandbox");
-    },
-    get webhookSignatureKey(): string {
-      return snap("square_webhook_signature_key");
-    },
-  },
-
-  // --- Stripe ---
-  stripe: withProperties(
-    {
-      get secretKey(): string {
-        return snap("stripe_secret_key");
-      },
-      get webhookEndpointId(): string {
-        return snap("stripe_webhook_endpoint_id");
-      },
-      get webhookSecret(): string {
-        return snap("stripe_webhook_secret");
-      },
-    },
-    providerKeyStatus("stripe_secret_key"),
-  ),
-
-  // --- SumUp ---
-  sumup: withProperties(
-    {
-      get apiKey(): string {
-        return snap("sumup_api_key");
-      },
-      get merchantCode(): string {
-        return snap("sumup_merchant_code");
-      },
-    },
-    providerKeyStatus("sumup_api_key"),
-  ),
-
   // --- Superuser ---
   get superuserChoice(): SuperuserChoice {
     const choice = snap("superuser_choice");
@@ -438,6 +272,7 @@ const settingsBase = {
   // -----------------------------------------------------------------------
   update: {
     ...stringAccessors.updaters,
+    ...paymentProviderAccessors.updaters,
     // --- Address lookup writes ---
     addressLookup: {
       apiKey: encryptedUpdate(CONFIG_KEYS.ADDRESS_LOOKUP_API_KEY),
@@ -463,11 +298,6 @@ const settingsBase = {
       CONFIG_KEYS.CALENDAR_FEEDS_GROUP_BY,
       "calendar_feeds_group_by",
     ) as (v: "attendees" | "listings") => Promise<void>,
-    clearPaymentProvider: async (): Promise<void> => {
-      await deleteRaw(CONFIG_KEYS.PAYMENT_PROVIDER);
-      data.payment_provider = null;
-      data.payment_provider_setting = null;
-    },
     contactFormEnabled: boolUpdate(
       CONFIG_KEYS.CONTACT_FORM_ENABLED,
       "contact_form_enabled",
@@ -508,63 +338,8 @@ const settingsBase = {
       CONFIG_KEYS.ORPHAN_PURGE_RETENTION,
       "orphan_purge_retention",
     ),
-    paymentProvider: async (v: PaymentProviderType): Promise<void> => {
-      await changePaymentProvider("activate", v);
-    },
-    paymentProviderAfterCredentialSave: async (
-      provider: PaymentProviderType,
-      activateFromMissing: boolean,
-    ): Promise<void> => {
-      await changePaymentProvider("credentials", provider, activateFromMissing);
-    },
-    recoverPaymentProvider: async (v: PaymentProviderType): Promise<void> => {
-      await changePaymentProvider("recover", v);
-    },
-    setPaymentProviderNone: async (): Promise<void> => {
-      await changePaymentProvider("disable");
-    },
     showPublicApi: boolUpdate(CONFIG_KEYS.SHOW_PUBLIC_API, "show_public_api"),
 
-    // --- Square writes ---
-    square: {
-      accessToken: encryptedUpdate(CONFIG_KEYS.SQUARE_ACCESS_TOKEN),
-      locationId: rawUpdate(
-        CONFIG_KEYS.SQUARE_LOCATION_ID,
-        "square_location_id",
-      ),
-      sandbox: boolUpdate(CONFIG_KEYS.SQUARE_SANDBOX, "square_sandbox"),
-      webhookSignatureKey: encryptedUpdate(
-        CONFIG_KEYS.SQUARE_WEBHOOK_SIGNATURE_KEY,
-      ),
-    },
-    // --- Stripe writes ---
-    stripe: {
-      configure: async (config: {
-        secretKey: string;
-        webhookSecret: string;
-        webhookEndpointId: string;
-      }): Promise<void> => {
-        // The API key and webhook pair belong to one Stripe account. Save all
-        // three together so a failed write leaves the prior credentials usable.
-        await writeRawBatch([
-          [CONFIG_KEYS.STRIPE_SECRET_KEY, await encrypt(config.secretKey)],
-          [
-            CONFIG_KEYS.STRIPE_WEBHOOK_SECRET,
-            await encrypt(config.webhookSecret),
-          ],
-          [CONFIG_KEYS.STRIPE_WEBHOOK_ENDPOINT_ID, config.webhookEndpointId],
-        ]);
-        data.stripe_secret_key = config.secretKey;
-        data.stripe_webhook_secret = config.webhookSecret;
-        data.stripe_webhook_endpoint_id = config.webhookEndpointId;
-      },
-      secretKey: encryptedUpdate(CONFIG_KEYS.STRIPE_SECRET_KEY),
-    },
-    // --- SumUp writes ---
-    sumup: {
-      apiKey: encryptedUpdate(CONFIG_KEYS.SUMUP_API_KEY),
-      merchantCode: plaintextUpdate(CONFIG_KEYS.SUMUP_MERCHANT_CODE),
-    },
     superuserChoice: plaintextUpdate(CONFIG_KEYS.SUPERUSER_CHOICE) as (
       v: SuperuserChoice,
     ) => Promise<void>,
@@ -590,4 +365,9 @@ const settingsBase = {
   withCurrentTask,
 };
 
-export const settings = withProperties(settingsBase, stringAccessors.getters);
+/** The namespace: the literal above, plus the two parts that build their own
+ *  getters. */
+export const settings = withProperties(
+  withProperties(settingsBase, paymentProviderAccessors.getters),
+  stringAccessors.getters,
+);
